ec2.py: More convenient grouping constructs
Put each EC2 machine in a group that matches its Name tag.  This way you
can refer to hosts with there FQDN just like hosts in your local data
center.

Support arbitrary grouping with an EC2 tag of "ansible_groups".  This
should be a comma delimited list of additional Ansible groups that the
host will be made a member of.